+++ src/config/index.js
@@ -1,5 +1,6 @@
 export { default as appInformation } from 'config/appInformation';
 export { default as englishContent } from 'config/languages/englishContent';
+export { default as italianContent } from 'config/languages/italianContent';
 export { default as navigationConfig } from 'config/navigationConfig';
 export { default as socialLinks } from 'config/socialLinks';
 export { default as spanishContent } from 'config/languages/spanishContent';