+++ src/config/languages/supportedLanguages.js
@@ -1,8 +1,9 @@
-import { englishContent, spanishContent } from 'config';
+import { englishContent, italianContent, spanishContent } from 'config';
 
 const supportedLanguages = {
     es: spanishContent,
     en: englishContent,
+    it: italianContent,
     default: englishContent
 };